refactor(provisa-ui): decompose GraphFrame and GraphPage into components/graph
Split the 3000+ line GraphFrame.tsx and 1500+ line GraphPage.tsx into a
cohesive components/graph module, each file under the 1000-line limit:

- GraphFrame split into GraphCanvas, Inspector, TableView, NodeContextMenu
  and pure helpers (cytoscape-types, graph-clusters, graph-export,
  graph-stylesheet, graph-model)
- GraphPage split into GraphSidebar, QueryBar, graph-context-menus,
  graph-persistence, graph-schema-types
- relocate the whole graph cluster from pages/ to components/graph to match
  the components/admin convention (GraphPage is the route; these are its parts)
- update test imports to the new graph-model location
